[Feat/#24] vc에서 view로만 전환
